+++ ohmg/frontend/svelte/src/lib/utils.js
@@ -52,22 +52,6 @@ export function projectionFromImageExtent(extent) {
 	});
 }
 
-export function submitPostRequest(url, headers, operation, payload, callback) {
-	const body = JSON.stringify({
-		"operation": operation,
-		"payload": payload,
-	});
-	fetch(url, {
-		method: 'POST',
-		headers: headers,
-		body: body,
-	})
-	.then(response => response.json())
-	.then(result => {
-		if (callback) {callback(result)}
-	});
-}
-
 export function makeTitilerXYZUrl (options) {
 	// options must be an object with the following properties:
 	// {
